reworked nn
